+++ changelog.md
@@ -1,3 +1,4 @@
 # Changelog
 
-Initial Commit
+**./classic/index.js**
+* Wrote demo code for 'ora-classic' package.